#916 商品用のハンドラから商品画像の処理を分割
