+++ clean-backend/internal/interface/http/router/router.go
@@ -19,6 +19,7 @@ func NewRouter(
 	creatorHandler *handler.CreatorHandler,
 	contactHandler *handler.ContactHandler,
 	productHandler *handler.ProductHandler,
+	productImageHandler *handler.ProductImageHandler,
 	userHandler *handler.UserHandler,
 	auth *middleware.AuthMiddleware,
 	admin *middleware.AdminMiddleware,
@@ -86,9 +87,9 @@ func NewRouter(
 	r.Handle("/api/product/{product_uuid}", requireAdmin(productHandler.Delete)).Methods(http.MethodDelete)
 
 	// product image
-	r.HandleFunc("/api/product_image/{product_image_uuid}/blob", productHandler.GetImageBlob).Methods(http.MethodGet)
-	r.Handle("/api/product/{product_uuid}/product_image", requireAdmin(productHandler.CreateImage)).Methods(http.MethodPost)
-	r.Handle("/api/product/{product_uuid}/product_image/{product_image_uuid}", requireAdmin(productHandler.DeleteImage)).Methods(http.MethodDelete)
+	r.HandleFunc("/api/product_image/{product_image_uuid}/blob", productImageHandler.GetBlob).Methods(http.MethodGet)
+	r.Handle("/api/product/{product_uuid}/product_image", requireAdmin(productImageHandler.Create)).Methods(http.MethodPost)
+	r.Handle("/api/product/{product_uuid}/product_image/{product_image_uuid}", requireAdmin(productImageHandler.Delete)).Methods(http.MethodDelete)
 
 	// contact
 	r.Handle("/api/contact", requireAdmin(contactHandler.List)).Methods(http.MethodGet)
